Explicitly typecast std::int64_t --> double for std::log2() argument
Also, use std:: namespace qualifier for this function

Hopefully fixes all type conversion warnings with Intel compiler
(unlike 498f8f31680e4f430530d5b5b2c9e12d334c3388)

Strangely, intel/17.0/64/17.0.5.239 on Perseus flags this warning, but
local macOS ICC version 19.0.1.144 does not.

+++ src/pgen/poisson.cpp
@@ -183,7 +183,10 @@ void Mesh::UserWorkAfterLoop(ParameterInput *pin) {
       std::int64_t zones = GetTotalCells();
       std::int64_t mb_zones = GetTotalCells()/nbtotal*nblist[Globals::my_rank];
       double zc_cpus = static_cast<double>(mb_zones*ncycle)/cpu_time;
-      double zc_cpus2 = static_cast<double>(mb_zones*log2(mb_zones)*ncycle)/cpu_time;
+      double zc_cpus2 =
+          static_cast<double>(mb_zones
+                              *static_cast<double>(std::log2(mb_zones))
+                              *ncycle)/cpu_time;
 
       if (Globals::my_rank == 0) {
         std::cout << "Timing Possison Solver                               " << std::endl;
